+++ src/cloudai/installer/slurm_installer.py
@@ -22,7 +22,7 @@ import toml
 from cloudai._core.base_installer import BaseInstaller
 from cloudai._core.system import System
 from cloudai._core.test_template import TestTemplate
-from cloudai.schema.system import SlurmSystem
+from cloudai.systems import SlurmSystem
 
 
 class SlurmInstaller(BaseInstaller):
