+++ src/cloudai/parser/system_parser/slurm_system_parser.py
@@ -16,7 +16,7 @@ import os
 from typing import Any, Dict, List
 
 from cloudai._core.base_system_parser import BaseSystemParser
-from cloudai.schema.system.slurm import SlurmNode, SlurmNodeState, SlurmSystem
+from cloudai.systems.slurm import SlurmNode, SlurmNodeState, SlurmSystem
 
 
 class SlurmSystemParser(BaseSystemParser):
