+++ src/cloudai/runner/slurm/slurm_runner.py
@@ -20,7 +20,7 @@ from cloudai._core.exceptions import JobIdRetrievalError
 from cloudai._core.system import System
 from cloudai._core.test import Test
 from cloudai._core.test_scenario import TestScenario
-from cloudai.schema.system import SlurmSystem
+from cloudai.systems import SlurmSystem
 from cloudai.util import CommandShell
 
 from .slurm_job import SlurmJob
