+++ src/cloudai/schema/test_template/jax_toolbox/slurm_command_gen_strategy.py
@@ -15,7 +15,7 @@
 import os
 from typing import Any, Dict, List
 
-from cloudai.schema.system.slurm.strategy import SlurmCommandGenStrategy
+from cloudai.systems.slurm.strategy import SlurmCommandGenStrategy
 
 
 class JaxToolboxSlurmCommandGenStrategy(SlurmCommandGenStrategy):
