+++ src/cloudai/schema/test_template/nemo_launcher/slurm_install_strategy.py
@@ -21,8 +21,8 @@ from concurrent.futures import ThreadPoolExecutor, as_completed
 from typing import Any, Dict, List
 
 from cloudai._core.system import System
-from cloudai.schema.system.slurm import SlurmNodeState, SlurmSystem
-from cloudai.schema.system.slurm.strategy import SlurmInstallStrategy
+from cloudai.systems.slurm import SlurmNodeState, SlurmSystem
+from cloudai.systems.slurm.strategy import SlurmInstallStrategy
 from cloudai.util import CommandShell
